% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitespace.R
\name{fct_whitespace_to_newline}
\alias{fct_whitespace_to_newline}
\title{Replaces all spaces with newlines for plotting}
\usage{
fct_whitespace_to_newline(x)
}
\arguments{
\item{x}{a tibble or data.frame to be used with \code{ggplot2}}
}
\description{
Uses regex to convert whitespaces (i.e. \code{\\s+}, not \code{" "})
into new lines (i.e. \code{\\n}). It applies to all factor/character
columns in a table.
}
