% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecdplot-package.R
\docType{data}
\name{oecd_colours}
\alias{oecd_colours}
\title{OECD Colours}
\format{
A data frame with 72 rows and 7 variables
}
\source{
OECD Chart Excel Add In
}
\usage{
oecd_colours
}
\description{
A data frame with 72 observations on the following 7 variables.
The variables are as follows:
}
\details{
\itemize{
  \item{\code{section}}{ section from the Add In (i.e. 'Full Colours')}
  \item{\code{palette}}{ the name of the palette (i.e. 'Dark Blue', 'Orange', etc.)}
  \item{\code{order}}{ the order of the colour (i.e. first, second, etc.)}
  \item{\code{red}}{ the amount of red of the colour (i.e. 131 in a 0-255 scale)}
  \item{\code{green}}{ the amount of green of the colour (i.e. 210 in a 0-255 scale)}
  \item{\code{blue}}{ the amount of blue of the colour (i.e. 227 in a 0-255 scale)}
  \item{\code{hex}}{ the hex colour for the colour (i.e. \code{#83D2E3})}
}
}
\keyword{datasets}
