% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{oecd_pal}
\alias{oecd_pal}
\title{OECD Colour Palettes}
\usage{
oecd_pal(alpha = 1, begin = 0, end = 1, direction = 1, option = "darkblue")
}
\arguments{
\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in [0,1] at which the color map begins.}

\item{end}{The (corrected) hue in [0,1] at which the color map ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default,
colors are ordered from darkest to lightest. If -1, the order of colors is
reversed.}

\item{option}{A character string indicating the color map option to use.
Options are available:
\itemize{
 \item "darkblue" (Standard)
 \item "orange"
 \item "darkgreen"
 \item "blue"
 \item "green"
 \item "pink"
 \item "purple"
 \item "red"
 \item "ppt"
}}
}
\description{
A wrapper function around \code{oecd_clrs} to
 turn it into a palette function compatible with
 \code{\link[ggplot2]{discrete_scale}}.
}
\details{
See \code{oecd_clrs} for more information on the color palettes.
}
\examples{
library(scales)
show_col(oecd_pal()(8), ncol = 8)

}
