% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{oecd_pie}
\alias{oecd_pie}
\title{Wrapper to simplify pie plots.}
\usage{
oecd_pie(
  data = NULL,
  x = NULL,
  xmin = NULL,
  xmax = NULL,
  y = NULL,
  ymin = NULL,
  ymax = NULL,
  colour = NULL,
  stacked = FALSE,
  size = NULL,
  facet = NULL,
  facet2 = NULL,
  group = NULL,
  text = NULL,
  palette = "darkblue",
  palette_na = "#7F7F7F",
  alpha = 1,
  title = NULL,
  subtitle = NULL,
  coord = NULL,
  x_breaks = NULL,
  x_expand = NULL,
  x_include = NULL,
  x_labels = NULL,
  x_limits = NULL,
  x_sec_axis = ggplot2::waiver(),
  x_title = NULL,
  x_trans = "identity",
  y_breaks = NULL,
  y_expand = NULL,
  y_include = NULL,
  y_labels = NULL,
  y_limits = NULL,
  y_sec_axis = ggplot2::waiver(),
  y_title = NULL,
  y_trans = "identity",
  col_breaks = NULL,
  col_include = NULL,
  col_intervals = NULL,
  col_labels = NULL,
  col_legend_place = "t",
  col_legend_ncol = NULL,
  col_legend_nrow = NULL,
  col_limits = NULL,
  col_title = NULL,
  col_continuous = "gradient",
  facet_labels = NULL,
  facet_ncol = NULL,
  facet_nrow = NULL,
  facet_scales = "fixed",
  facet_space = "fixed",
  caption = NULL,
  theme = oecdplot::theme_oecd(),
  trend = FALSE,
  band = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{Unquoted x aesthetic variable.}

\item{xmin}{Unquoted xmin aesthetic variable.}

\item{xmax}{Unquoted xmin aesthetic variable.}

\item{y}{Unquoted y aesthetic variable.}

\item{ymin}{Unquoted ymin aesthetic variable. Defaults to "ymin" and it's
computed from "y" if it's not present.}

\item{ymax}{Unquoted ymin aesthetic variable. Defaults to "ymin" and it's
computed from "y" if it's not present.}

\item{colour}{Unquoted colour/fill aesthetic variable.}

\item{stacked}{Logical value to stack the columns.}

\item{size}{Unquoted size aesthetic variable.}

\item{facet}{Unquoted facet aesthetic variable.}

\item{facet2}{Unquoted second facet variable for a facet grid of facet by facet2 variables.}

\item{group}{Unquoted group aesthetic variable.}

\item{text}{Unquoted text aesthetic variable, which can be used in combination with plotly::ggplotly(., tooltip = "text").}

\item{palette}{Palette to use. A character vector of hex codes (or names). The default is to use "darkblue". See the documentation for `oecd_clrs()` to see all the options.}

\item{palette_na}{Colour to use for NA values. A character vector of a hex code (or name).}

\item{alpha}{Opacity. A number between 0 and 1 (the default is 1).}

\item{title}{Title string.}

\item{subtitle}{Subtitle string.}

\item{coord}{Coordinate system.}

\item{x_breaks}{A function that takes the limits as input (e.g. scales::breaks_pretty()), or a vector of breaks.}

\item{x_expand}{Padding to the limits with the ggplot2::expansion function, or a vector of length 2 (e.g. c(0, 0)).}

\item{x_include}{For a numeric or date variable, any values that the scale should include (e.g. 0).}

\item{x_labels}{A function that takes the breaks as inputs (e.g. scales::label_comma()), or a vector of labels.}

\item{x_limits}{A vector of length 2 to determine the limits of the axis.}

\item{x_sec_axis}{A secondary axis specified by the ggplot2::sec_axis or ggplot2::dup_axis function.}

\item{x_title}{Axis title string. Defaults to converting to sentence case with spaces. Use "" for no title.}

\item{x_trans}{For a numeric variable, a transformation object (e.g. "log10").}

\item{y_breaks}{A function that takes the limits as input (e.g. scales::breaks_pretty()), or a vector of breaks.}

\item{y_expand}{Padding to the limits with the ggplot2::expansion function, or a vector of length 2 (e.g. c(0, 0)).}

\item{y_include}{For a numeric or date variable, any values that the scale should include (e.g. 0).}

\item{y_labels}{A function that takes the breaks as inputs (e.g. scales::label_comma()), or a vector of labels.}

\item{y_limits}{A vector of length 2 to determine the limits of the axis.}

\item{y_sec_axis}{A secondary axis specified by the ggplot2::sec_axis or ggplot2::dup_axis function.}

\item{y_title}{Axis title string. Defaults to converting to sentence case with spaces. Use "" for no title.}

\item{y_trans}{For a numeric variable, a transformation object (e.g. "log10").}

\item{col_breaks}{A function that takes the limits as input (e.g. scales::breaks_pretty()), or a vector of breaks.}

\item{col_include}{For a numeric or date variable, any values that the scale should include (e.g. 0).}

\item{col_intervals}{A function to cut or chop the numeric variable into intervals (e.g. ~ santoku::chop_mean_sd(.x, drop = FALSE)).}

\item{col_labels}{A function that takes the breaks as inputs (e.g. scales::label_comma()), or a vector of labels. Note this does not affect where col_intervals is not NULL.}

\item{col_legend_place}{The place for the legend. "b" for bottom, "r" for right, "t" for top (default), or "l" for left.}

\item{col_legend_ncol}{The number of columns for the legend elements.}

\item{col_legend_nrow}{The number of rows for the legend elements.}

\item{col_limits}{A vector to determine the limits of the axis.}

\item{col_title}{Axis title string. Defaults to converting to sentence case with spaces. Use "" for no title.}

\item{col_continuous}{Type of colouring for a continuous variable. Either "gradient" or "steps". Defaults to "steps".}

\item{facet_labels}{A function that takes the breaks as inputs (e.g. scales::label_comma()), or a named vector of labels (e.g. c(value = "label", ...)).}

\item{facet_ncol}{The number of columns of facetted plots.}

\item{facet_nrow}{The number of rows of facetted plots.}

\item{facet_scales}{Whether facet scales should be "fixed" across facets, "free" in both directions, or free in just one direction (i.e. "free_x" or "free_y"). Defaults to "fixed".}

\item{facet_space}{Whether facet space should be "fixed" across facets, "free" to be proportional in both directions, or free to be proportional in just one direction (i.e. "free_x" or "free_y"). Only applies Where facet2 is provided and facet_scales are not fixed. Defaults to "fixed".}

\item{caption}{Caption title string.}

\item{theme}{A ggplot2 theme. The default is to use `theme_oecd()`.}

\item{trend}{Add a trend line (line plot only). Defaults to `FALSE`.}

\item{band}{Add a confidence band (line plot only). Defaults to `FALSE`.}

\item{...}{Other arguments passed to the relevant ggplot2::geom_* function.}
}
\value{
A ggplot object.
}
\description{
Create a pie plot with a wrapper around the ggplot2::geom_col function.
}
\examples{
\dontrun{
# THIS TYPE OF PLOT IS NOT RECOMMENDED AS IT PRESENTS IMPORTANT
# PERCEPTUAL ISSUES

pta_chl <- data.frame(
  country = "CHL",
  category = c("Protected", "Non-Protected"),
  pct_pta = c(20.5, 79.5)
)

oecd_pie(pta_chl, x = country, y = pct_pta, colour = category, group = category)
}

}
