% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{oecd_watermark}
\alias{oecd_watermark}
\title{Save 'ggplot' Objects According to PAC Rules}
\usage{
oecd_watermark(x = 20, y = 20, text = "Draft")
}
\arguments{
\item{x}{X-axis coordinate for the watermark (defaults to 20)}

\item{y}{Y-axis coordinate for the watermark (defaults to 20)}

\item{text}{The text for the watermark (defaults to "Draft")}
}
\description{
Uses predefined sizes in centimeters (cm) to export graphic objects. Besides
saving EMF and PNG files, optionally creates a spreadsheet with Statlinks.
}
\examples{
\dontrun{
library(ggplot2)

p <- ggplot(pta) +
  geom_col(
    aes(x = country, y = pct_pta, fill = category),
    position = "dodge2"
  )

p + oecd_watermark()
}

}
