% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecdplot-package.R
\docType{package}
\name{oecdplot-package}
\alias{oecdplot}
\alias{oecdplot-package}
\title{oecdplot: Wrappers for 'ggplot2'}
\description{
Provides themes and scales (i.e. color and fill) according to PAC style rules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://algobank.oecd.org:4430/r-oecd-graphs/oecdplot}
  \item Report bugs at \url{https://algobank.oecd.org:4430/r-oecd-graphs/oecdplot/-/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mv.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item Ana Moreno Monroy \email{Ana.MorenoMonroy@oecd.org} (\href{https://orcid.org/0000-0002-8282-4067}{ORCID}) [contributor]
  \item Bethany Millar-Powell \email{Bethany.Millar-Powell@oecd.org} [contributor]
  \item Harriet Goers \email{Harriet.GOERS@oecd.org} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [contributor]
  \item Maria Paula Caldas \email{MariaPaula.CALDAS@oecd.org} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [contributor]
  \item  OECD [data contributor, copyright holder, funder]
}

}
\keyword{internal}
