% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecdplot-package.R
\docType{data}
\name{pta}
\alias{pta}
\title{Protected Terrestrial Areas}
\format{
A data frame with 99 rows and 4 variables
}
\source{
OECD Stats
}
\usage{
pta
}
\description{
Regional disparities in protected terrestrial areas (2017) expressed as
protected terrestrial areas as a percentage of the total area for large
regions. See the Stat link https://doi.org/10.1787/888934190039.
}
\details{
\itemize{
  \item{\code{country}}{ ISO-3 code of for the countries}
  \item{\code{region}}{ Geographical region for the observations}
  \item{\code{category}}{ Category for the observations (minimum, country
  value or maximum)}
  \item{\code{pct_pta}}{ Protected terrestrial areas a percentage of the
  total area (large regions)}
}
}
\examples{
pta
}
\keyword{datasets}
