% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_oecd_chart}
\alias{save_oecd_chart}
\title{Save 'ggplot' Objects According to PAC Rules}
\usage{
save_oecd_chart(
  file_name = "graph name",
  plot = last_plot(),
  folder = "publication/",
  size = "1/2",
  dpi = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_note = NULL,
  plot_source = NULL,
  statlink_create = FALSE,
  statlink_dataframe = NULL,
  statlink_filename = NULL,
  box = FALSE,
  custom_size = NULL,
  format = c("png", "emf"),
  emfPlus = NULL,
  emfPlusFontToPath = NULL
)
}
\arguments{
\item{file_name}{Name for the file}

\item{plot}{Graphic object to export (defaults to the last plot in the
Viewer)}

\item{folder}{Golder where the plot will be saved}

\item{size}{Chart size according to official PAC dimensions in centimeters\tabular{lll}{
   Parameter \tab Width \tab Height \cr
   1/3 \tab 15.42 \tab 7.08 \cr
   1/3 \tab 15.42 \tab 9.66 \cr
   2/3 \tab 15.42 \tab 11.93 \cr
   1/3 (2 graphs in the same row) \tab 7.54 \tab 7.08 \cr
   1/2 (2 graphs in the same row) \tab 7.54 \tab 9.66 \cr
   2/3 (2 graphs in the same row) \tab 7.54 \tab 11.93 \cr
   Full footnotes \tab 13.22 \tab 10.49 \cr
   Full page \tab 13.22 \tab 20.8 \cr
   Slides \tab 26 \tab 12.2 \cr
   Half slides \tab 14 \tab 12.2 \cr
}}

\item{dpi}{Chart dots per inch (DPI). The default is NULL.}

\item{plot_title}{Title of the plot}

\item{plot_subtitle}{Subtitle of the plot}

\item{plot_note}{Note for the plot}

\item{plot_source}{Source of the plot}

\item{statlink_create}{\code{TRUE} or \code{FALSE} (default)}

\item{statlink_dataframe}{Data used to create figure that will be added to
Statlink}

\item{statlink_filename}{the filename with no extension}

\item{box}{Logical parameter to rescale plots for boxes (reduced plot width)}

\item{custom_size}{(Optional) A vector such as \code{c(20.8, 13.22, 49)} (i.e.
20.8 cm width, 13.22 cm height, data starts in row 49 in the exported Excel)}

\item{format}{a string or vector indicating the formats. The options are
"png", "emf", "ps" and the default is NULL, which saves as PNG and EMF}

\item{emfPlus}{Totally optional parameter for EMF rendering}

\item{emfPlusFontToPath}{Totally optional parameter for EMF rendering}
}
\description{
Uses predefined sizes in centimeters (cm) to export graphic objects. Besides
saving EMF and PNG files, optionally creates a spreadsheet with Statlinks.
}
\examples{
\dontrun{
save_oecd_chart(
  file_name = "fig1",
  plot = f1,
  folder = "dev/pac_demo/",
  size = "2/3",
  plot_title = "Life Expectancy in 2002 and 2007",
  plot_subtitle = "Source: Gapminder",
  plot_note = "Note: Random sample of countries",
  plot_source = "R 4.1",
  statlink_create = T,
  statlink_dataframe = d1,
  statlink_filename = "f1"
)
}
}
