% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R, R/fill.R
\name{scale_colour_oecd_d}
\alias{scale_colour_oecd_d}
\alias{scale_colour_oecd_c}
\alias{scale_fill_oecd_d}
\alias{scale_fill_oecd_c}
\title{OECD Colour Scales}
\usage{
scale_colour_oecd_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "darkblue",
  aesthetics = "colour"
)

scale_colour_oecd_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "darkblue",
  values = NULL,
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "colour"
)

scale_fill_oecd_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "darkblue",
  aesthetics = "fill"
)

scale_fill_oecd_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "darkblue",
  values = NULL,
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Other arguments passed on to [ggplot2::discrete_scale()],
[ggplot2::continuous_scale()], or [ggplot2::binned_scale()] to control name, limits, breaks,
  labels and so forth.}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in [0,1] at which the color map begins.}

\item{end}{The (corrected) hue in [0,1] at which the color map ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default,
colors are ordered from darkest to lightest. If -1, the order of colors is
reversed.}

\item{option}{A character string indicating the color map option to use.
Options are available:
\itemize{
 \item "darkblue" (Standard)
 \item "orange"
 \item "darkgreen"
 \item "blue"
 \item "green"
 \item "pink"
 \item "purple"
 \item "red"
 \item "ppt"
}}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the `colour` and `fill` aesthetics at the
same time, via `aesthetics = c("colour", "fill")`.}

\item{values}{if colours should not be evenly positioned along the gradient this
vector gives the position (between 0 and 1) for each colour in the colours vector.}

\item{na.value}{If \code{na.translate = TRUE}, what aesthetic value should the
missing values be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}
}
\description{
The `oecd` scales provide official colour maps for the different OECD's
teams. They are based on the official XML especifications for the Excel
add-in.
}
\examples{
\dontrun{
library(ggplot2)

load_oecd_fonts()

p <- ggplot(pta) +
  geom_col(
    aes(x = country, y = pct_pta, fill = category),
    position = "dodge2"
  )

# both fill and fill direction are optional
p + scale_fill_oecd_d(option = "darkgreen", direction = -1)
}
}
\concept{colour scales}
