% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_oecd}
\alias{theme_oecd}
\title{'ggplot2' Theme Compatible with PAC Rules}
\usage{
theme_oecd(
  base_size = 7.5,
  base_family = "Arial Narrow",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  base_background = "white",
  base_x_axis_angle = 45,
  base_col_width = 0.7
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family}

\item{base_line_size}{Base size forline elements}

\item{base_rect_size}{Base size for rect elements}

\item{base_background}{Base background color (white or gray/grey)}

\item{base_x_axis_angle}{Base angle of the x-axis text (default 45)}

\item{base_col_width}{Base angle of the column width (default 0.7)}
}
\description{
Complete theme which controls all non-data display. You can use \code{theme()}
in a chained expression to tweak the display for this current theme.
}
\examples{
\dontrun{
library(ggplot2)

load_oecd_fonts()

p <- ggplot(pta) +
  geom_col(
    aes(x = country, y = pct_pta, fill = category),
    position = "dodge2"
  )

# both fill and fill direction are optional
p + theme_oecd()
}
}
